pyproject.toml
[project]
name = "agentsphere-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "alembic==1.17.2",
    "asyncpg==0.30.0",
    "ddgs==9.9.1",
    "duckduckgo-search==8.1.1",
    "email-validator>=2.3.0",
    "google-api-python-client==2.187.0",
    "google-auth-httplib2==0.2.1",
    "google-auth-oauthlib==1.2.3",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "langchain==1.0.7",
    "langchain-community==0.4.1",
    "langchain-core==1.0.5",
    "langchain-experimental>=0.4.0",
    "langchain-google-genai==0.0.1",
    "langchain-groq==1.0.1",
    "langchain-openai==1.0.3",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.3",
    "langgraph-checkpoint==3.0.1",
    "langgraph-checkpoint-postgres==3.0.1",
    "langgraph-prebuilt==1.0.4",
    "langgraph-sdk==0.2.9",
    "langgraph-supervisor==0.0.30",
    "langsmith==0.4.49",
    "mcp>=1.1.2",
    "numexpr>=2.14.1",
    "numpy==2.3.5",
    "openai==2.8.0",
    "groq==0.34.1",
    "playwright==1.56.0",
    "psycopg2-binary==2.9.11",
    "psycopg-binary==3.2.12",
    "python-dotenv>=1.0.0",
    "python-multipart==0.0.20",
    "python-pptx>=0.6.21",
    "requests==2.32.5",
    "sse-starlette==3.0.3",
    "starlette==0.50.0",
    "sqlalchemy==2.0.44",
    "tiktoken==0.12.0",
    "tqdm==4.67.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.38.0",
    "xlsxwriter==3.2.9",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
    "google-generativeai==0.3.2",
    "pydantic==2.11.1",
    "pillow>=8.0.0",
    "fonttools>=4.0.0",
    "anthropic>=0.3.0",
]
